api-gateway/requirements.txt
# API Gateway Dependencies

fastapi
uvicorn[standard]
grpcio
grpcio-tools
protobuf
pydantic[email]
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
prometheus-client
httpx

# Development tools
flake8
black
isort
pre-commit

# Testing
pytest
pytest-asyncio
pytest-mock
httpx  # For testing FastAPI endpoints
